requirements.txt
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만
djangorestframework>=3.14.0,<4.0.0
psycopg2-binary>=2.9.9,<3.0
